move cidrtoddn to utils
